LayoutTests/http/tests/appcache/abort-cache-onchecking-manifest-404-expected.txt

CONSOLE MESSAGE: line 1: ApplicationCache is deprecated. Please use ServiceWorkers instead.
CONSOLE MESSAGE: line 13: Application Cache download process was aborted.
This tests that download process was aborted after checking event and before update failed.
SUCCESS
